pyproject.toml
# 核心：使用 Poetry 原生配置，不再混合 PEP 621
[tool.poetry]
# 必填元数据（description 非空，解决核心错误）
name = "app"
version = "0.1.0"
description = "Llama3.2 project with FastAPI, Qdrant and Ollama integration"  # 非空描述
authors = ["tangyangsheng <29817019911@qq.com>"]  # 简化格式，也可保留字典格式
license = "MIT"
readme = "README.md"
keywords = ["fastapi", "llama-index", "ollama", "embedding"]
# 项目结构配置
packages = [{ include = "app", from = "src" }]
include = ["src/"]

# Poetry 依赖（保留原有依赖，无变动）
[tool.poetry.dependencies]
python = "^3.13.9"
qdrant-client = "^1.16.1"
uvicorn = "^0.38.0"
fastapi = "^0.123.8"
python-multipart = "^0.0.20"
pydantic-settings = "^2.12.0"
llama-index-core = ">=0.10.1,<0.10.86"
llama-index-llms-ollama = ">=0.1.11"
llama-index-embeddings-huggingface = ">=0.1.7"
llama-index-readers-file = ">=0.1.28"
transformers = ">=4.41.0"
torch = ">=2.3.0"
sentence-transformers = ">=2.7.0"
injector = "^0.23.0"
python-dotenv = ">=0.21.0"
pyyaml = "^6.0.2"
retry-async = "^0.1.4"
llama-index-vector-stores-qdrant = "^0.1.4"

# 构建系统（保留）
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# 镜像源：标记清华源为默认，提升依赖安装速度（可选，但推荐）
[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary" 

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "supplemental"